[sui-fork] Move StartArgs into args.rs and fold fork.rs into lib.rs
Match the layout of the indexer-alt crates: clap argument structs live in
an args module, and the service types (ForkNode, ForkAdmin, and their
result types) are defined directly in lib.rs rather than in a module
named after the service. fork.rs is gone; the root re-exports are
unchanged apart from the types now being defined at the root.
